requirements.txt
requests~=2.31.0
cryptography~=41.0.3
opencv-python~=4.11.0.86
python-dotenv~=0.21.0